First pass at Navigation module
In this attempt we are not using any modules, which feels wrong, but
this feels like  a step in the right direction.

In app/app.js we require our Navigation module, and two views (which
requiring views here also feels wrong).  We then set both of these views
to monitor changes to the Navigation module's 'ci' variable (short for
current_index).

Then in our app/router.js file we create a few new
routes.

Our Navigation model has a validate function which does not let us set
the 'ci' variable to a step that does not exist or is greater than
'max_page'.

To test this:
  1) Click on the 'page 2' link, you should get a console message
     stating that 'You are not allowed to go that far forward'
  2) Click on the 'next >>' link.  Note that the 'page 2' link becomes
     active.
  3) Click on the 'page 1' link.  Note that you are allowed to go
     backwards.
  4) Click on the 'page 2' link again.  Note that you are now allowed to
     move this far forward.
  5) Repeat onwards.

Known Issues:
  - Once you are already on the 'next >>' link (#/next), you can not
    click the 'next >>' link again to move to the next step with out
    first clicking on another page.
  - Once you've already advanced through the system, if you go backwards
    and then click the 'next >>' link again, it does not move you
    forward.